+++ applications/fishing-map/src/features/app/app.selectors.ts
@@ -21,7 +21,10 @@ import {
   WorkspaceState,
   WorkspaceStateProperty,
 } from 'types'
-import { selectDataviewInstancesMerged } from 'features/dataviews/dataviews.selectors'
+import {
+  selectActiveVesselsDataviews,
+  selectDataviewInstancesMerged,
+} from 'features/dataviews/dataviews.selectors'
 
 export const selectViewport = createSelector(
   [selectUrlViewport, selectWorkspaceViewport],
@@ -97,9 +100,9 @@ export const selectTimebarEvents = createSelector(
 )
 
 export const selectTimebarGraph = createSelector(
-  [selectWorkspaceStateProperty('timebarGraph')],
-  (timebarGraph): TimebarGraphs => {
-    return timebarGraph
+  [selectWorkspaceStateProperty('timebarGraph'), selectActiveVesselsDataviews],
+  (timebarGraph, vessels): TimebarGraphs => {
+    return vessels && vessels.length ? timebarGraph : TimebarGraphs.None
   }
 )