+++ applications/fishing-map/src/features/datasets/datasets.hook.ts
@@ -11,6 +11,7 @@ import {
 import {
   getContextDataviewInstance,
   getEnvironmentDataviewInstance,
+  getUserTrackDataviewInstance,
 } from 'features/dataviews/dataviews.utils'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import { useAppDispatch } from 'features/app/app.hooks'
@@ -34,23 +35,31 @@ import {
 
 const DATASET_REFRESH_TIMEOUT = 10000
 
-export const useNewDatasetConnect = () => {
+export const useAddDataviewFromDatasetToWorkspace = () => {
   const contextDataviews = useSelector(selectContextAreasDataviews)
   const activityDataviews = useSelector(selectActivityDataviews)
   const enviromentalDataviews = useSelector(selectEnvironmentalDataviews)
   const { upsertDataviewInstance } = useDataviewInstancesConnect()
 
-  const addNewDatasetToWorkspace = useCallback(
+  const addDataviewFromDatasetToWorkspace = useCallback(
     (dataset: Dataset) => {
       let dataviewInstance
       if (dataset.category === DatasetCategory.Context) {
         const usedColors = contextDataviews?.flatMap((dataview) => dataview.config?.color || [])
         dataviewInstance = getContextDataviewInstance(dataset.id, usedColors)
-      } else {
+      } else if (
+        dataset.category === DatasetCategory.Environment &&
+        dataset.configuration?.geometryType === 'polygons'
+      ) {
         const usedRamps = [...(enviromentalDataviews || []), ...(activityDataviews || [])].flatMap(
           (dataview) => dataview.config?.colorRamp || []
         )
         dataviewInstance = getEnvironmentDataviewInstance(dataset.id, usedRamps)
+      } else if (
+        dataset.category === DatasetCategory.Environment &&
+        dataset.configuration?.geometryType === 'tracks'
+      ) {
+        dataviewInstance = getUserTrackDataviewInstance(dataset)
       }
       if (dataviewInstance) {
         upsertDataviewInstance(dataviewInstance)
@@ -59,7 +68,7 @@ export const useNewDatasetConnect = () => {
     [contextDataviews, enviromentalDataviews, activityDataviews, upsertDataviewInstance]
   )
 
-  return { addNewDatasetToWorkspace }
+  return { addDataviewFromDatasetToWorkspace }
 }
 
 export const useDatasetModalConnect = () => {