+++ applications/fishing-map/src/features/workspace/context-areas/ContextAreaLayerPanel.tsx
@@ -2,17 +2,18 @@ import React, { useState } from 'react'
 import cx from 'classnames'
 import { useTranslation } from 'react-i18next'
 import { DatasetTypes, DatasetStatus } from '@globalfishingwatch/api-types'
-import { Switch, IconButton, Tooltip, ColorBar } from '@globalfishingwatch/ui-components'
-import {
-  ColorBarOption,
-  TrackColorBarOptions,
-} from '@globalfishingwatch/ui-components/dist/color-bar'
+import { Tooltip } from '@globalfishingwatch/ui-components'
+import { ColorBarOption } from '@globalfishingwatch/ui-components/dist/color-bar'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
 import styles from 'features/workspace/shared/LayerPanel.module.css'
 import { useDataviewInstancesConnect } from 'features/workspace/workspace.hook'
 import { useAutoRefreshImportingDataset } from 'features/datasets/datasets.hook'
-import ExpandedContainer from 'features/workspace/shared/ExpandedContainer'
 import DatasetNotFound from '../shared/DatasetNotFound'
+import Color from '../common/Color'
+import LayerSwitch from '../common/LayerSwitch'
+import InfoError from '../common/InfoError'
+import Remove from '../common/Remove'
+import Title from '../common/Title'
 
 type LayerPanelProps = {
   dataview: UrlDataviewInstance
@@ -20,22 +21,10 @@ type LayerPanelProps = {
 
 function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
   const { t } = useTranslation()
-  const { upsertDataviewInstance, deleteDataviewInstance } = useDataviewInstancesConnect()
+  const { upsertDataviewInstance } = useDataviewInstancesConnect()
   const [colorOpen, setColorOpen] = useState(false)
 
   const layerActive = dataview?.config?.visible ?? true
-  const onToggleLayerActive = () => {
-    upsertDataviewInstance({
-      id: dataview.id,
-      config: {
-        visible: !layerActive,
-      },
-    })
-  }
-
-  const onRemoveClick = () => {
-    deleteDataviewInstance(dataview.id)
-  }
 
   const changeColor = (color: ColorBarOption) => {
     upsertDataviewInstance({
@@ -66,10 +55,14 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
   const title = isUserLayer
     ? dataset?.name || dataset?.id
     : t(`datasets:${dataset?.id}.name` as any)
+
   const TitleComponent = (
-    <h3 className={cx(styles.name, { [styles.active]: layerActive })} onClick={onToggleLayerActive}>
-      {title}
-    </h3>
+    <Title
+      title={title}
+      className={styles.name}
+      classNameActive={styles.active}
+      dataview={dataview}
+    />
   )
 
   const datasetImporting = dataset.status === DatasetStatus.Importing
@@ -94,14 +87,11 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
       })}
     >
       <div className={styles.header}>
-        <Switch
+        <LayerSwitch
           disabled={datasetError}
           active={layerActive}
-          onClick={onToggleLayerActive}
-          tooltip={t('layer.toggleVisibility', 'Toggle layer visibility')}
-          tooltipPlacement="top"
           className={styles.switch}
-          color={dataview.config?.color}
+          dataview={dataview}
         />
         {title && title.length > 30 ? (
           <Tooltip content={title}>{TitleComponent}</Tooltip>
@@ -109,48 +99,22 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
           TitleComponent
         )}
         <div className={cx('print-hidden', styles.actions, { [styles.active]: layerActive })}>
-          <IconButton
-            icon={datasetError ? 'warning' : 'info'}
-            type={datasetError ? 'warning' : 'default'}
-            size="small"
+          <InfoError
+            error={datasetError}
             loading={datasetImporting}
-            className={styles.actionButton}
             tooltip={infoTooltip}
-            tooltipPlacement="top"
+            className={styles.actionButton}
           />
           {layerActive && (
-            <ExpandedContainer
-              visible={colorOpen}
+            <Color
+              dataview={dataview}
+              open={colorOpen}
+              onColorClick={changeColor}
+              onToggleClick={onToggleColorOpen}
               onClickOutside={closeExpandedContainer}
-              component={
-                <ColorBar
-                  colorBarOptions={TrackColorBarOptions}
-                  selectedColor={dataview.config?.color}
-                  onColorClick={changeColor}
-                />
-              }
-            >
-              <IconButton
-                icon={colorOpen ? 'color-picker' : 'color-picker-filled'}
-                size="small"
-                style={colorOpen ? {} : { color: dataview.config?.color }}
-                tooltip={t('layer.color_change', 'Change color')}
-                tooltipPlacement="top"
-                onClick={onToggleColorOpen}
-                className={cx(styles.actionButton)}
-              />
-            </ExpandedContainer>
-          )}
-          {isUserLayer && (
-            <IconButton
-              icon="delete"
-              size="small"
-              tooltip={t('layer.remove', 'Remove layer')}
-              tooltipPlacement="top"
-              onClick={onRemoveClick}
-              className={cx(styles.actionButton)}
             />
           )}
+          {isUserLayer && <Remove className={cx(styles.actionButton)} dataview={dataview} />}
         </div>
       </div>
     </div>