+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.38
+
+### Patch Changes
+
+- Updated dependencies [0ec2156f]
+  - @globalfishingwatch/dataviews-client@6.3.0
+  - @globalfishingwatch/layer-composer@5.3.0
+  - @globalfishingwatch/timebar@1.5.0
+  - @globalfishingwatch/api-client@1.6.9
+
 ## 1.1.37
 
 ### Patch Changes