+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @globalfishingwatch/react-hooks
 
+## 6.1.0
+
+### Minor Changes
+
+- 0ec2156f: User tracks
+
+### Patch Changes
+
+- Updated dependencies [0ec2156f]
+  - @globalfishingwatch/dataviews-client@6.3.0
+  - @globalfishingwatch/layer-composer@5.3.0
+  - @globalfishingwatch/ui-components@2.19.0
+  - @globalfishingwatch/fourwings-aggregate@3.1.1
+  - @globalfishingwatch/api-client@1.6.9
+
 ## 6.0.10
 
 ### Patch Changes